\name{aut}
\alias{aut}
\docType{data}
\title{
Administrative Boundaries for Austria
}
\description{
High precision geospatial data for Austria's administrative Boundaries.
}
\usage{data("aut")}
\references{
https://www.data.gv.at/katalog/dataset/51bdc6dc-25ae-41de-b8f3-938f9056af62
}
\examples{
data(aut)

names(aut)
}
\keyword{datasets}
