\name{sbg}
\alias{sbg}
\docType{data}

\title{
Salzburg Mapping Grid
}

\description{
Grid covering the state of Salzburg (Austria).
}

\usage{data("sbg")}

\references{
Niklfeld, H. (1978): Grundfeldschlüssel für die Kartierung der Flora Mitteleuropas, (Zentralstelle für Florenkartierung, Inst. Bot. Univ. Wien). Wien.
}

\examples{
data(sbg)

plot(sbg)
names(sbg)
}

\keyword{datasets}

