\name{sabotag-package}
\alias{sabotag-package}
\alias{sabotag}
\docType{package}
\title{
\packageTitle{sabotag}
}
\description{
\packageDescription{sabotag}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sabotag}
\packageIndices{sabotag}

An R-package providing tools for the Salzburger Botanische Arbeitsgemeinschaft (SaBotAg)

}
\author{
\packageAuthor{sabotag}

Maintainer: \packageMaintainer{sabotag}
}

\keyword{ package }
\keyword{ spatial }

\examples{
data(sbg)
}
